Fast_API/requirements.txt
python-dotenv
ipykernel
langchain-groq
fastapi
langserve
sse_starlette